Extracted `fn apply_instrumentation{,_point}` to `crate::point::apply`.
